Break config class out into its own file
In preparation for making ProviderConfig into a general locally saved config, there's nothing inherently linked to the providers other than that's where we initially ran into needing them.
